Moved SW specific tests from manager to a new file
* Tests for ServiceWorkerManager.ts should test only the usage from a page.
* The new context/sw tests should be for the ServiceWorker itself.
